feat: Update import/export names for modular files
